Automanager 2 pushbots 1 scorebot Trussdoor broken
I did not commit this until I had almost everything added. I have 2 working Pushbots, and a Scoreing bot. I do not have a demo yet. The field was modeled around the CAD file for the field from the place you buy it. I did not use the real CAD file for it because it had 20 million triangles it's self. So I make a replica in blender. There is a Autonomous manager that runs the game for 30 seconds until triggring a C# event for. GL to all!